Updated readme and the Broad Overviews
Short description: 

```
This paper argues that it is dangerous to think of machine learning as offering quick wins coming for free. The authors  focus on the aspect of technical debt, and conclude that it is easy to build in massive ongoing maintenance costs at the system level
when applying machine learning. Several machine learning specific risk factors are also high lighted and design patterns to be avoided or refactored where possible.

```

This article describe the challenges that machine learning introduces to software development(SW) and puts a lot of attention to which aspects that are different then classical SW development. I believe it is relevant and also very often referred to in conferences and applied ML and ML operations. Reading this article gives the reader a better understanding of common mistakes and  lessons from the industry related to technical debt in machine learning systems. Compare to the other articles this article takes the standpoint from operations, maintenance and and technical debt.