Adding article: Close the AI Accountability Gap
Short description -- Proposes a deliberate structured process for performing a risk assessment (internal audit) on an AI system.
Why is it relevant? -- Provides operational guidance for governance on machine learning applications
What is the lesson learned from reading the article? -- Many useful lessons, including practical hints on what type of artifacts to produce, what steps to take, how to compose a team, but also on the limitations of an internal audit
How does it add new insights over the articles already listed? -- The other articles in the Governance section call for risk assessments and audits, but do not describe how to conduct them