Update readme.md
Short description: The article presents a summary of technical and non-technical solutions for bias in ML. It also advances the idea that technical debiasing techniques are not sufficient to remove the negative impact ML can have.  

Why is it relevant? The idea of thinking beyond technical debiasing technique is unique and relevant for ML project governance.

What is the lesson learned from reading the article? One of the main lessons learned is that technical debiasing techniques are not sufficient.

How does it add new insights over the articles already listed? By promoting the idea of thinking beyond technical solutions, i.e., to more general governance objectives.